Fix warning in hazard_pointer_test.cc

+++ containers_cpp/test/hazard_pointer_test.cc
@@ -74,7 +74,7 @@ int* IntObjectTestPool::Allocate() {
 }
 
 void IntObjectTestPool::Release(int* object_pointer) {
-  int cell = object_pointer - simplePoolObjects;
+  int cell = (int) (object_pointer - simplePoolObjects);
   simplePool[cell].Store(FREE_MARKER);
 }
